--- gear4j-extra/src/main/java/club/gclmit/gear4j/extra/notify/Dingtalk/channel/Button.java
@@ -1,33 +1,0 @@
-package club.gclmit.gear4j.extra.notify.Dingtalk.channel;
-
-/**
- * ActionCard Button 按钮
- *
- * @author gclm
- */
-public class Button {
-
-    private String title;
-    private String actionURL;
-
-    public Button(String title, String actionURL) {
-        this.title = title;
-        this.actionURL = actionURL;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getActionURL() {
-        return actionURL;
-    }
-
-    public void setActionURL(String actionURL) {
-        this.actionURL = actionURL;
-    }
-}
