+++ gear4j-extra/src/main/java/club/gclmit/gear4j/extra/notify/dingtalk/channel/DingtalkChannel.java
@@ -1,9 +1,9 @@
-package club.gclmit.gear4j.extra.notify.Dingtalk.channel;
+package club.gclmit.gear4j.extra.notify.dingtalk.channel;
 
 
 import java.util.List;
 
-import club.gclmit.gear4j.extra.notify.Dingtalk.DingtalkBot;
+import club.gclmit.gear4j.extra.notify.dingtalk.DingtalkBot;
 
 /**
  * 封装钉钉消息通道
