+++ gear4j-extra/src/main/java/club/gclmit/gear4j/extra/notify/dingtalk/channel/Text.java
@@ -1,10 +1,10 @@
-package club.gclmit.gear4j.extra.notify.Dingtalk.channel;
+package club.gclmit.gear4j.extra.notify.dingtalk.channel;
 
 import java.util.List;
 
 import com.alibaba.fastjson.JSONObject;
 
-import club.gclmit.gear4j.extra.notify.Dingtalk.DingtalkBot;
+import club.gclmit.gear4j.extra.notify.dingtalk.DingtalkBot;
 
 /**
  * Text 消息
